backend/requirements.txt
openai>=1.50.0
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
rich>=13.0.0
